Hue/pyproject.toml
[project]
name = "hue-async-poc"
version = "0.1.0"
description = "Local-only Philips Hue room controller (prototype + scalable design)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "requests>=2.31.0",
  "python-dotenv>=1.0.0",
  "pydantic-settings>=2.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
